+++ pygqlc/MutationParser.py
@@ -1,5 +1,10 @@
 import re
 
+"""
+The purpose of this module is to prepare a graphql transaction, such as a query
+ or mutation, to be able to carry out a batch of them.
+"""
+
 mutation_regex = r'^\s*mutation\s*(\s+[a-zA-Z_]+[a-zA-Z_0-9]?)?\s*(\(\s*(((\$[a-zA-Z_]+[a-zA-Z_0-9]?)\s*:\s*([a-zA-Z_]+[a-zA-Z_0-9]?!?)\s*)+\s*)\))?\s*{\s*((.\s*)+\s*})\s*}'
 rgx_groups = {
   'full_doc': 0,
@@ -8,7 +13,25 @@ rgx_groups = {
   'content': 7,
 }
 class MutationParser:
+  """This is the MutationParser class, it can parse a graphql instructions.
+  """
   def __init__(self, gql_doc):
+    """Constructor of the MutationParser object.
+
+    Args:
+        gql_doc (string): Graphql instructions
+
+    Attributes:
+        re (module): Regexp module.
+        regex (regex): Transaction Regex.
+        match (Match Object): Match object of the pattern and doc. Defaults to
+          None.
+        isValid (boolean): Checks if the gql_doc is valid. Defaults to False.
+        full_doc (string): All Graphql instructions.
+        alias (string): Name of the transaction.
+        variables (string): Variables of the transaction.
+        content (content): The transaction content.
+    """
     self.re = re
     self.regex = mutation_regex
     self.gql_doc = gql_doc
@@ -20,6 +43,11 @@ class MutationParser:
     self.content = None
 
   def parse(self):
+    """This fuction parses and validates the transaction instructions.3
+
+    Returns:
+        (boolean): Returns if the parsed doc was succesful.
+    """
     # ! First, remove variable definitions:
     doc = self.gql_doc
     var_end = doc.find('{')
@@ -34,6 +62,12 @@ class MutationParser:
     return False
 
   def validate(self):
+    """This function checks if the transaction instructions has match with the
+     transaction regex.
+
+    Returns:
+        (boolean): Returns if it is valid or not.
+    """
     match = self.re.match(self.regex, self.gql_doc)
     self.isValid = match is not None
     if self.isValid:
@@ -41,6 +75,14 @@ class MutationParser:
     return self.isValid
   
   def format_value(self, value):
+    """This function formats document's values.
+
+    Args:
+        value (any): Value that want to format.
+
+    Returns:
+        (string): Returns a formated string.
+    """
     if type(value) == str:
       return f'"{value}"'
     elif type(value) == bool: