+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [2.0.0] - 2020-10-20
 ### Added
-- Add a Singleton metaclass to the helper_modules module
+- Add Documentation with sphinx
+### Changed
+- Change Singleton implementation from decorator to metaclass
 ### Fixed
 - Dependencies have been set to be installed from certain compatible version
 ### Removed